feat: MOVE-2144 - Change content verifier provider
Replace bc package's PGP content verifier builder provider with
corresponsiding provider in jcajce package. The .bc. packages are
needed either where the JCA/JCE is not available on the JVM's
profile, or where we might be deploying remotely and unable to
manage the policy files for the JCE. We require JCE for installing the
integrasjonspunkt application.